src/main/webapp/app/entities/event-log-book/update/event-log-book-update.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of, Subject, from } from 'rxjs';

import { EventLogBookService } from '../service/event-log-book.service';
import { IEventLogBook, EventLogBook } from '../event-log-book.model';

import { IUser } from 'app/entities/user/user.model';
import { UserService } from 'app/entities/user/user.service';

import { EventLogBookUpdateComponent } from './event-log-book-update.component';

describe('EventLogBook Management Update Component', () => {
  let comp: EventLogBookUpdateComponent;
  let fixture: ComponentFixture<EventLogBookUpdateComponent>;
  let activatedRoute: ActivatedRoute;
  let eventLogBookService: EventLogBookService;
  let userService: UserService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      declarations: [EventLogBookUpdateComponent],
      providers: [
        FormBuilder,
        {
          provide: ActivatedRoute,
          useValue: {
            params: from([{}]),
          },
        },
      ],
    })
      .overrideTemplate(EventLogBookUpdateComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(EventLogBookUpdateComponent);
    activatedRoute = TestBed.inject(ActivatedRoute);
    eventLogBookService = TestBed.inject(EventLogBookService);
    userService = TestBed.inject(UserService);

    comp = fixture.componentInstance;
  });

  describe('ngOnInit', () => {
    it('Should call User query and add missing value', () => {
      const eventLogBook: IEventLogBook = { id: 456 };
      const user: IUser = { id: 90609 };
      eventLogBook.user = user;

      const userCollection: IUser[] = [{ id: 63030 }];
      jest.spyOn(userService, 'query').mockReturnValue(of(new HttpResponse({ body: userCollection })));
      const additionalUsers = [user];
      const expectedCollection: IUser[] = [...additionalUsers, ...userCollection];
      jest.spyOn(userService, 'addUserToCollectionIfMissing').mockReturnValue(expectedCollection);

      activatedRoute.data = of({ eventLogBook });
      comp.ngOnInit();

      expect(userService.query).toHaveBeenCalled();
      expect(userService.addUserToCollectionIfMissing).toHaveBeenCalledWith(userCollection, ...additionalUsers);
      expect(comp.usersSharedCollection).toEqual(expectedCollection);
    });

    it('Should update editForm', () => {
      const eventLogBook: IEventLogBook = { id: 456 };
      const user: IUser = { id: 18202 };
      eventLogBook.user = user;

      activatedRoute.data = of({ eventLogBook });
      comp.ngOnInit();

      expect(comp.editForm.value).toEqual(expect.objectContaining(eventLogBook));
      expect(comp.usersSharedCollection).toContain(user);
    });
  });

  describe('save', () => {
    it('Should call update service on save for existing entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<EventLogBook>>();
      const eventLogBook = { id: 123 };
      jest.spyOn(eventLogBookService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ eventLogBook });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: eventLogBook }));
      saveSubject.complete();

      // THEN
      expect(comp.previousState).toHaveBeenCalled();
      expect(eventLogBookService.update).toHaveBeenCalledWith(eventLogBook);
      expect(comp.isSaving).toEqual(false);
    });

    it('Should call create service on save for new entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<EventLogBook>>();
      const eventLogBook = new EventLogBook();
      jest.spyOn(eventLogBookService, 'create').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ eventLogBook });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: eventLogBook }));
      saveSubject.complete();

      // THEN
      expect(eventLogBookService.create).toHaveBeenCalledWith(eventLogBook);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).toHaveBeenCalled();
    });

    it('Should set isSaving to false on error', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<EventLogBook>>();
      const eventLogBook = { id: 123 };
      jest.spyOn(eventLogBookService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ eventLogBook });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.error('This is an error!');

      // THEN
      expect(eventLogBookService.update).toHaveBeenCalledWith(eventLogBook);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).not.toHaveBeenCalled();
    });
  });

  describe('Tracking relationships identifiers', () => {
    describe('trackUserById', () => {
      it('Should return tracked User primary key', () => {
        const entity = { id: 123 };
        const trackResult = comp.trackUserById(0, entity);
        expect(trackResult).toEqual(entity.id);
      });
    });
  });
});
